requirements.txt
pymongo==3.9.0
python-chess==0.29.0
pyspark==2.4.4
matplotlib==3.1.2
pandas==0.25.3